+++ v3/crates/metadata-resolve/src/stages/boolean_expressions/mod.rs
@@ -15,9 +15,9 @@ use crate::Qualified;
 
 pub use types::{
     BooleanExpressionComparableRelationship, BooleanExpressionGraphqlConfig,
-    BooleanExpressionGraphqlFieldConfig, BooleanExpressionTypes, BooleanExpressionsOutput,
-    ComparisonExpressionInfo, IncludeLogicalOperators, ObjectComparisonExpressionInfo,
-    ResolvedObjectBooleanExpressionType,
+    BooleanExpressionGraphqlFieldConfig, BooleanExpressionIssue, BooleanExpressionTypes,
+    BooleanExpressionsOutput, ComparisonExpressionInfo, IncludeLogicalOperators,
+    ObjectComparisonExpressionInfo, ResolvedObjectBooleanExpressionType,
 };
 
 pub fn resolve(
@@ -31,6 +31,7 @@ pub fn resolve(
     object_types: &type_permissions::ObjectTypesWithPermissions,
 ) -> Result<BooleanExpressionsOutput, BooleanExpressionError> {
     let mut raw_boolean_expression_types = BTreeMap::new();
+    let mut issues = vec![];
 
     // TODO: make sure we are adding new types here, we are almost certainly not doing this atm
     let graphql_types = existing_graphql_types.clone();
@@ -56,7 +57,10 @@ pub fn resolve(
         if let BooleanExpressionOperand::Object(boolean_expression_object_operand) =
             &boolean_expression_type.operand
         {
-            let object_boolean_expression_type = object::resolve_object_boolean_expression_type(
+            let object::ObjectBooleanExpressionTypeOutput {
+                object_boolean_expression: object_boolean_expression_type,
+                issues: object_issues,
+            } = object::resolve_object_boolean_expression_type(
                 boolean_expression_type_name,
                 boolean_expression_object_operand,
                 &boolean_expression_type.logical_operators,
@@ -67,6 +71,8 @@ pub fn resolve(
                 &raw_boolean_expression_types,
                 graphql_config,
             )?;
+
+            issues.extend(object_issues);
 
             boolean_expression_object_types.insert(
                 boolean_expression_type_name.clone(),
@@ -81,5 +87,6 @@ pub fn resolve(
             scalars: boolean_expression_scalar_types.clone(),
         },
         graphql_types,
+        issues,
     })
 }