+++ v3/crates/metadata-resolve/src/types/warning.rs
@@ -1,4 +1,4 @@
-use crate::stages::object_boolean_expressions;
+use crate::stages::{boolean_expressions, object_boolean_expressions};
 
 /// Warnings for the user raised during metadata generation
 /// These are things that don't break the build, but may do so in future
@@ -8,4 +8,6 @@ pub enum Warning {
     ObjectBooleanExpressionWarning(
         #[from] object_boolean_expressions::ObjectBooleanExpressionWarning,
     ),
+    #[error("{0}")]
+    BooleanExpressionIssue(#[from] boolean_expressions::BooleanExpressionIssue),
 }